+++ src/bc_coverage.cpp
@@ -89,7 +89,7 @@ main (int argc, char* argv[]) {
     string out_prefix;
 
     char bc_delim = ':';
-    uint8_t bc_col = 7;
+    size_t bc_col = 7;
     string bc_tag;
 
     size_t min_mapq = 0;
