+++ src/bc_tss_enrichment.cpp
@@ -104,7 +104,7 @@ main (int argc, char* argv[]) {
     size_t max_frag_len = 1000;
 
     char bc_delim = ':';
-    uint8_t bc_col = 7;
+    size_t bc_col = 7;
     string bc_tag;
 
     size_t min_mapq = 0;
