+++ src/pseudobulk_fragment_metagene.cpp
@@ -88,7 +88,7 @@ main (int argc, char* argv[]) {
     size_t n_divisions = 100;
 
     char bc_delim = ':';
-    uint8_t bc_col = 7;
+    size_t bc_col = 7;
     string bc_tag;
 
     size_t min_frag_len = 0;
